fix: Additional type annotation fixes for mypy compliance
- Fix async generator type in BaseStreamProcessor interface
- Add missing return type annotations across all modules
- Fix variable shadowing in api.py classify_data endpoint
- Add proper type annotations for feature extractors
- Fix dataclass .dict() calls with asdict() function
- Add missing type annotations for inner functions
- Initialize _previous_features attribute in SeizureFeatureExtractor
- Cast numpy float returns to Python float for type safety
- Add missing Optional import
- Apply Black formatting

Note: Some mypy errors remain due to design decisions:
- Specific result types don't inherit from ClassificationResult base
- Numpy type inference issues with generic arrays
These require architectural changes beyond the scope of linting fixes.

+++ neural-engine/src/classification/classifiers/mental_state.py
@@ -382,7 +382,7 @@ class MentalStateClassifier(BaseClassifier):
 
     def _sigmoid(self, x: float) -> float:
         """Sigmoid activation function"""
-        return 1.0 / (1.0 + np.exp(-x))
+        return float(1.0 / (1.0 + np.exp(-x)))
 
     async def load_model(self, model_path: str) -> None:
         """Load pre-trained model"""
